Fix search/pagination bug, fix issue with task number after page refresh
